edit data format

+++ dynamixel_workbench_toolbox/include/dynamixel_workbench_toolbox/dynamixel_driver.h
@@ -101,7 +101,7 @@ class DynamixelDriver
   bool reboot(uint8_t id, const char **log = NULL);
   bool reset(uint8_t id, const char **log = NULL);
 
-  bool writeRegister(uint8_t id, uint16_t address, uint8_t length, uint8_t* data, const char **log = NULL);
+  bool writeRegister(uint8_t id, uint16_t address, uint16_t length, uint8_t* data, const char **log = NULL);
 
   bool writeRegister(uint8_t id, const char *item_name, uint8_t data, const char **log = NULL);
   bool writeRegister(uint8_t id, const char *item_name, uint16_t data, const char **log = NULL);
@@ -113,7 +113,7 @@ class DynamixelDriver
   bool writeOnlyRegister(uint8_t id, const char *item_name, uint16_t data, const char **log = NULL);
   bool writeOnlyRegister(uint8_t id, const char *item_name, uint32_t data, const char **log = NULL);  
 
-  bool readRegister(uint8_t id, uint16_t address, uint16_t length, uint8_t *data, const char **log = NULL);
+  bool readRegister(uint8_t id, uint16_t address, uint16_t length, uint32_t *data, const char **log = NULL);
 
   bool readRegister(uint8_t id, const char *item_name, uint8_t *data, const char **log = NULL);
   bool readRegister(uint8_t id, const char *item_name, uint16_t *data, const char **log = NULL);
