+++ dynamixel_workbench_toolbox/include/dynamixel_workbench_toolbox/dynamixel_item.h
@@ -77,7 +77,7 @@ typedef struct
   const char *item_name;
   uint16_t    address;
   uint8_t	    item_name_length;
-  uint8_t     data_length;
+  uint16_t    data_length;
 } ControlItem;
 
 typedef struct
